+++ src/main/java/org/kochab/simulatedannealing/Scheduler.java
@@ -1,5 +1,5 @@
 package org.kochab.simulatedannealing;
 
 public interface Scheduler {
-    double getTemperature(int steps);
+    double getTemperature(long steps);
 }
